+++ services/functiongraph/v2/model/model_month_used.go
@@ -12,7 +12,7 @@ type MonthUsed struct {
 	Date *string `json:"date,omitempty"`
 
 	// 使用量
-	Value *float32 `json:"value,omitempty"`
+	Value *int32 `json:"value,omitempty"`
 }
 
 func (o MonthUsed) String() string {
